+++ CHANGES.rst
@@ -3,7 +3,8 @@ _____
 
 *Features*
 
-- radices : Removed `RadixBase` class, `BasedReal` implementations now declared on their own
+- radices : Remove `RadixBase` class, `BasedReal` implementations now declared on their own
+- tables : Implement `models` features in `HTable`, `HTable.from_model` and models attributes
 
 0.5.2
 _____